AD-HOC refactor (variables): Renamed a variable after ansible conventions
Normally, Ansible variables are of the form "node_exporter_exporters".
Previously, the author had fashioned the variables after the "helm"
project form Kubernetes, which did not match the pattern nor code style
of other roles.

This has been refactored for ease of reading, and to reduce the
"surprise". Further, the variable is now an array rather than a hard
coded list, better reflecting the nature of the data.