AD-HOC refactor (*): Rename variables to the Ansible standard
Previously variables in the defaults had bee named after a style that is
common in the tooling around "helm", from the Kubernetes project.
However, this style is not common in the Ansible world; preferring
instead to prefix the variable with the role, and then the variable
purpose.

This commit brings all variables into this standard.