AD-HOC fix (Binary): Allow unarchive to work from local machines
Previously, this role had only been executed by Ansible on the same
machine as the node-exporter is being installed in. This meant that
ansible would both install items on the "remote" in the correct
location, and find them on the "local" in that same location. The
ansible unarchive task by default looks for resources on the local
machine, so when this module is being used "remotely" it fails with the
inability to find the archive.

This commit adds the required flag to the unarchive task such that it
finds the archive where expected